reorganize actions and cacheHeaders
